requirements.txt
streamlit
transformers
sentence-transformers
faiss-cpu
#pyngrok
#nltk
symspellpy